src/doc-qa/pipeline/requirements.txt
kfp>=2.7.0
google-cloud-aiplatform>=1.38.0
